+++ test/fun_with_flags/simple_store_test.exs
@@ -123,6 +123,47 @@ defmodule FunWithFlags.SimpleStoreTest do
   end
 
 
+  describe "all_flags() returns all the flags" do
+    test "with no saved flags it returns an empty list" do
+      clear_redis_test_db()
+      assert [] = SimpleStore.all_flags()
+    end
+
+    test "with saved flags in returns a list of flags" do
+      clear_redis_test_db()
+
+      name1 = unique_atom()
+      g_1a = Gate.new(:boolean, false)
+      g_1b = Gate.new(:actor, "the actor", true)
+      g_1c = Gate.new(:group, :horses, true)
+      SimpleStore.put(name1, g_1a)
+      SimpleStore.put(name1, g_1b)
+      SimpleStore.put(name1, g_1c)
+
+      name2 = unique_atom()
+      g_2a = Gate.new(:boolean, false)
+      g_2b = Gate.new(:actor, "another actor", true)
+      SimpleStore.put(name2, g_2a)
+      SimpleStore.put(name2, g_2b)
+
+      name3 = unique_atom()
+      g_3a = Gate.new(:boolean, true)
+      SimpleStore.put(name3, g_3a)
+
+      result = SimpleStore.all_flags()
+      assert 3 = length(result)
+
+      for flag <- [
+        %Flag{name: name1, gates: [g_1a, g_1b, g_1c]},
+        %Flag{name: name2, gates: [g_2a, g_2b]},
+        %Flag{name: name3, gates: [g_3a]}
+      ] do
+        assert flag in result
+      end
+    end
+  end
+
+
   describe "integration: enable and disable with the top-level API" do
     test "looking up a disabled flag" do
       name = unique_atom()